made responsive
summary: at 750px and below, connect 4 is on top and my profile info is below it. also the board and tiles sizes change to be smaller. another div containing text is also centered, div starting with working to grow as a front-end developer.

test plan: ran npm start and inspected how the layout looked on different devices with google and inspect. They look good in my opinion.