Sent boardArr, currentPlayer and currentColumns to client. Get these variables in client and save in state.
Summary: The above. Now when I click on the game board on the client, after a refresh, the game board colours show in the client.

Test Plan: Ran both npm start and npm run dev. Encountered issue. Between lines 28 - 35, on Board.js, with the useEffect, having an empty and the game colours show after a refresh. I want the effect to run after every click on the game board. If I add boardArr, currentPlayer and currentColumns as dependency, it generate an infinite loop.