put variables showNewGameButton, winner and gameOver on the server, changes functions
summary: I adjusted the code and functions so that the variables of winner, gameOver and showNewGameButton are on the server and adjusted the functions so that the game works properly.

test plan: Ran npm start and npm run dev, games seems to work. Deployed main branch on heroku and the game state seems to work properly now on heroku!